215. Implementing the create order method

+++ Core/Entities/OrderAggregate/ProductItemOrdered.cs
@@ -11,14 +11,14 @@ namespace Core.Entities.OrderAggregate
         {
         }
 
-        public ProductItemOrdered(int productItemId, string productName, string pictureUrl)
+        public ProductItemOrdered(long productItemId, string productName, string pictureUrl)
         {
             ProductItemId = productItemId;
             ProductName = productName;
             PictureUrl = pictureUrl;
         }
 
-        public int ProductItemId { get; set; }
+        public long ProductItemId { get; set; }
         public string ProductName { get; set; }
         public string PictureUrl { get; set; }
     }
